pyproject.toml
[project]
name = "basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.0.3",
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    "langgraph>=1.0.0",
    "python-dotenv>=1.2.1",
    "grandalf>=0.8",
    "pillow>=12.0.0",
    "langchainhub>=0.1.21",
    "langchain-core>=1.0.5",
    "langsmith>=0.4.43",
    "ipython>=9.7.0",
    "langchain-tavily>=0.2.13",
]
authors = [
    { name="victor", email="xonme888@gmail.com" },
]
license = { file="LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
